requirements.txt
flask
python-chess
stockfish
torch
numpy